+++ premier-design/features/documents-content/ui/document-image/DocumentImage.tsx
@@ -4,6 +4,8 @@ import styles from './DocumentImage.module.css';
 import {FC} from "react";
 import {DocumentImageProps} from "@features/documents-content/interface/DocumentImage.props";
 
+const DOCUMENT_IMAGE_SIZES = '(max-width: 600px) 100vw, (max-width: 1440px) 60vw, 1935px';
+
 const DocumentImage: FC<DocumentImageProps> = ({alt, src}) => (
     <div className={styles.content__image}>
         <Image
@@ -13,11 +15,7 @@ const DocumentImage: FC<DocumentImageProps> = ({alt, src}) => (
             placeholder='empty'
             width={2000}
             height={160}
-            sizes="
-                (max-width: 600px) 100vw,
-                (max-width: 1440px) 60vw,
-                1935px
-            "
+            sizes={DOCUMENT_IMAGE_SIZES}
             className={styles.document_image}
         />
     </div>